requirements.txt
kubernetes
openstacksdk
oslo.limit
